refactor register intercom
